+++ src/views/Playersite.js
@@ -98,7 +98,7 @@ export default function Playersite(props) {
 		const value = event.target.value;
 		const settings = $settings;
 
-		if (['playersite', 'header', 'share', 'support'].includes(what)) {
+		if (['playersite', 'header', 'share', 'support', 'chromecast'].includes(what)) {
 			settings[what] = !settings[what];
 		} else {
 			settings[what] = value;
@@ -418,6 +418,14 @@ export default function Playersite(props) {
 										checked={$settings.share}
 										disabled={!$settings.playersite}
 										onChange={handleChange('share')}
+									/>
+								</Grid>
+								<Grid item xs={12}>
+									<Checkbox
+										label={<Trans>Chromecast</Trans>}
+										checked={$settings.chromecast}
+										disabled={!$settings.playersite}
+										onChange={handleChange('chromecast')}
 									/>
 								</Grid>
 								<Grid item xs={12}>