+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 #### v1.1.0 > v1.2.0
 
+-   Add Chromecast support (thx badincite, [#10](https://github.com/datarhei/restreamer-ui/pull/10))
 -   Add stream distribution across multiple internal servers
 -   Add SRT settings
 -   Add HLS version selection (Dwaynarang, Electra Player compatibility)